+++ netbox/circuits/filters.py
@@ -125,6 +125,7 @@ class CircuitFilter(CustomFieldFilterSet, django_filters.FilterSet):
             Q(cid__icontains=value) |
             Q(terminations__xconnect_id__icontains=value) |
             Q(terminations__pp_info__icontains=value) |
+            Q(terminations__description__icontains=value) |
             Q(description__icontains=value) |
             Q(comments__icontains=value)
         ).distinct()
@@ -160,5 +161,6 @@ class CircuitTerminationFilter(django_filters.FilterSet):
         return queryset.filter(
             Q(circuit__cid__icontains=value) |
             Q(xconnect_id__icontains=value) |
-            Q(pp_info__icontains=value)
+            Q(pp_info__icontains=value) |
+            Q(description__icontains=value)
         ).distinct()