+++ CHANGELOG.md
@@ -14,6 +14,7 @@ NetBox now supports modeling physical cables for console, power, and interface c
 ## Enhancements
 
 * [#450](https://github.com/digitalocean/netbox/issues/450) - Added `outer_width` and `outer_depth` fields to rack model
+* [#867](https://github.com/digitalocean/netbox/issues/867) - Added `description` field to circuit terminations
 * [#1444](https://github.com/digitalocean/netbox/issues/1444) - Added an `asset_tag` field for racks
 * [#2000](https://github.com/digitalocean/netbox/issues/2000) - Dropped support for Python 2
 * [#2104](https://github.com/digitalocean/netbox/issues/2104) - Added a `status` field for racks
@@ -37,6 +38,7 @@ NetBox now supports modeling physical cables for console, power, and interface c
     * `power_ports`
     * `pass_through_ports`
 * The field `interface_ordering` has been removed from the DeviceType serializer
+* Added a `description` field to the CircuitTermination serializer
 
 ---